Refactor Gulp
